+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - TypeScript definitions related to `Payload`, `PayloadType`, and `PayloadParameters` are also exported.
 - New `ValidationError` error.
 - The `Payload` class now supports serialization and deserialization to bytes / Buffer.
+- The `PayloadArray` class is now part of the countertop. It has a slightly modified API.
 
 ## [0.4.1] - 2022-06-08
 ### Changed